Removed Error caused by front end shenanigans
I was getting an error in my console for putting in the wrong class name and...I just don't like red text showing up there so I sought out a way to fix it. I figured it out and removed it on my project so I figured why not put it here. 

I forked the newest version and put the change in. Normally I say something in console when I test for errors and find them so at first I did that. It might've been helpful to a developer maybe but using this but they should be focused no making sure their css classes are spelled right and that they are using the right ones anyways it's not hard. This way it just behaves more like something like font awesome or whatever. 

...Been using this a couple days and I really like it! It's different from what I'm used to with font-awesome and